在Mac上运行遭遇的一堆问题
多运行几遍，每一次运行都按第一个按钮，全部崩溃。有趣的是，每一次崩溃，
崩溃的地点都不一样，有的在FT崩溃，有的在GL崩溃，有的在renderString
崩溃。最终发生错误的直接原因也不一样，有的是在libsystem_platform.dylib
里的_platform_memmove$VARIANT$Haswell + 77中因为EXC_BAD_ACCESS崩溃，
有的是在经过libGL、GLEngine、libGPUSupportMercury、
AppleIntelBDWGraphicsGLDriver、libsystem_malloc、libsystem_kernel几层
库后，因为"incorrect checksum for freed object - object was probably
modified after being freed."错误崩溃。我今天晚上花了两个小时试图探查
其原因，但是无果。如果谁还有闲暇时间，欢迎投入到艰苦卓绝的debug工作。

调用栈回溯信息可以找dlaboratory@126.com用电子邮件索要。

+++ source/World.cpp
@@ -615,9 +615,9 @@ void updateblock(int x, int y, int z, bool updated, int depth)
         }
         br = brts[maxbrightness];
         if (blks[maxbrightness] == block(Blocks::WATER))
-            br = max(br - 2, BRIGHTNESSMIN);
+            br = max(br - 2, (int)BRIGHTNESSMIN);
         else
-            br = max(br - 1, BRIGHTNESSMIN);
+            br = max(br - 1, (int)BRIGHTNESSMIN);
         if (skylighted && br < skylight)
             br = skylight;
         br = max(br, BRIGHTNESSMIN);
